build script added

+++ OS/day5/fact.c
@@ -1,14 +1,14 @@
 #include <stdio.h>
 
-int factr(int n) {
+double factr(double n) {
 	if(n == 0 || n == 1)
 		return 1;
 	else
 		return factr(n-1)*n;		
 }
 
-int facti(int n) {
-	int result = 1;
+double facti(double n) {
+	double result = 1;
 	while(n>0){
 		result *= n;
 		n--;
